pyproject.toml
[project]
name = "stock2"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "finance-datareader>=0.9.110",
    "google-api-python-client>=2.193.0",
    "google-auth>=2.49.1",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "google-genai>=1.68.0",
    "gspread>=6.2.1",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "requests>=2.32.5",
]

[tool.pyright]
extraPaths = ["./src"]
typeCheckingMode = "basic"
